Modularized cmdUtil
